Add a readme.

+++ setup.sh
@@ -7,9 +7,13 @@ then
     cd ..
 else
     echo "Cloning the COVID-19 JHU repository."
-    rm -r COVID-19
+    if [ -d "./COVID-19" ]
+    then
+        rm -r COVID-19
+    fi
     git clone git@github.com:CSSEGISandData/COVID-19.git
 fi
+
 # Make sure that South Korea is correctly named.
 echo "Making sure South Korea named correctly."
 find ./COVID-19 \( -type d -name .git -prune \) -o -type f -print0 | xargs -0 sed -i 's/"Korea, South"/South Korea/g'
